pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gh-issue-workflow"
version = "0.1.0"
description = "Standalone multi-repo GitHub issue stage-label workflow automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
gh-issue-workflow = "gh_issue_workflow.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
